chg: Moved JoeParser class to make it reachable from expansion & import modules
