+++ misp_modules/lib/__init__.py
@@ -1,0 +1,1 @@
+all = ['joe_parser']
